🚚 rename `show_node` module to `nodes`
